feat: Update dependencies and improve onboarding UI
- Updated `dots_indicator` to version 4.0.1 for improved functionality.
- Updated `go_router` to version 16.0.0 for enhanced routing capabilities.
- Updated `google_sign_in` to version 7.1.0 for better integration.
- Updated `sentry_flutter` to version 9.2.0 for improved error tracking.
- Updated `sign_in_with_apple` to version 7.0.1 for better compatibility.
- Updated `flutter_lints` to version 6.0.0 for improved linting rules.
- Updated `sentry_dart_plugin` to version 3.1.0 for enhanced features.
- Updated various other dependencies in `pubspec.lock` for stability.
- Fixed onboarding UI by converting `_currentPage` to double for `DotsIndicator`.
- Added `jni` plugin to both Linux and Windows generated plugin files.

+++ lib/features/on_boarding/presentation/views/widgets/on_boarding_view_body.dart
@@ -60,7 +60,7 @@ class _OnBoardingViewBodyState extends State<OnBoardingViewBody> {
               // ─── Dots ───
               DotsIndicator(
                 dotsCount: 3,
-                position: _currentPage,
+                position: _currentPage.toDouble(),
                 decorator: const DotsDecorator(
                   activeColor: AppColors.primary,
                   color: AppColors.lightPrimary,
